backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests>=2.28.0,<3.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
aiohttp

# 可选：AI服务相关依赖
openai==1.3.0
tencentcloud-sdk-python==3.0.896

# 开发依赖
pytest==7.4.2
black==23.9.1
flake8==6.1.0